Move base event in a common package not related to a domain
