+++ domains/books/_data_access_interfaces.py
@@ -4,7 +4,7 @@ from typing import Any, List, Protocol, Tuple, Union
 from sqlalchemy_bind_manager.repository import SortDirection
 
 from domains.books.models import BookModel
-from domains.cloudevent_base import BaseEvent
+from domains.common.cloudevent_base import BaseEvent
 
 
 class BookRepositoryInterface(Protocol):
