+++ domains/common/README.md
@@ -1,0 +1,6 @@
+# Common module
+
+This module contains logic that needs to be consistent across all the domains but doesn't specifically
+belong to one of them. E.g. 
+
+* Base classes (base model, base event, etc.)
